splitted gui from main.py
